+++ apps/meteor/app/lib/server/functions/getFullUserData.ts
@@ -7,7 +7,7 @@ import { settings } from '../../../settings/server';
 
 const logger = new Logger('getFullUserData');
 
-const defaultFields = {
+export const defaultFields = {
 	name: 1,
 	username: 1,
 	nickname: 1,
@@ -19,12 +19,13 @@ const defaultFields = {
 	reason: 1,
 	statusText: 1,
 	avatarETag: 1,
-	extension: 1,
 	federated: 1,
 	statusLivechat: 1,
+	abacAttributes: 1,
+	freeSwitchExtension: 1,
 } as const;
 
-const fullFields = {
+export const fullFields = {
 	emails: 1,
 	phone: 1,
 	statusConnection: 1,
@@ -35,7 +36,6 @@ const fullFields = {
 	requirePasswordChangeReason: 1,
 	roles: 1,
 	importIds: 1,
-	freeSwitchExtension: 1,
 } as const;
 
 let publicCustomFields: Record<string, 0 | 1> = {};
@@ -86,7 +86,6 @@ export async function getFullUserDataByIdOrUsernameOrImportId(
 		(searchType === 'username' && searchValue === caller.username) ||
 		(searchType === 'importId' && caller.importIds?.includes(searchValue));
 	const canViewAllInfo = !!myself || (await hasPermissionAsync(userId, 'view-full-other-user-info'));
-	const canViewExtension = !!myself || (await hasPermissionAsync(userId, 'view-user-voip-extension'));
 
 	// Only search for importId if the user has permission to view the import id
 	if (searchType === 'importId' && !canViewAllInfo) {
@@ -98,7 +97,6 @@ export async function getFullUserDataByIdOrUsernameOrImportId(
 	const options = {
 		projection: {
 			...fields,
-			...(canViewExtension && { freeSwitchExtension: 1 }),
 			...(myself && { services: 1 }),
 		},
 	};